# Add a pass for FastExp conversion