@asaadaldien 2020-11-18
> et's move it to Conve

Yeah having an LLVMIR -> LLVMIR pass is confusing move it to LLVMToLLVM directory