@asaadaldien 2020-11-18
> hat would let us see the IR and verify it without needing to rely on the execution tests? or is it too messy?

I think I can add a test to verify `llvm.intr.exp -> correct_sequance_of_llvm_ops` but its very verbose so the only test that is added in this PR is a numerical check test.